Adding amortized graveyard hashing
