+++ Makefile
@@ -1,10 +1,18 @@
 TARGETS=test test_threadsafe test_pc bm hm_churn test_runner
 
+FEATURE_FLAGS= 
 ifdef T
-  FEATURE_FLAGS= -DQF_TOMBSTONE
-else
-  FEATURE_FLAGS= 
+		FEATURE_FLAGS:=$(FEATURE_FLAGS) -DQF_TOMBSTONE
 endif
+
+ifdef REBUILD_BY_CLEAR
+		FEATURE_FLAGS:=$(FEATURE_FLAGS) -DREBUILD_BY_CLEAR
+endif
+
+ifdef REBUILD_AMORTIZED_GRAVEYARD
+		FEATURE_FLAGS:=$(FEATURE_FLAGS) -DREBUILD_AMORTIZED_GRAVEYARD
+endif
+
 
 ifdef D
 	DEBUG=-g -DDEBUG=1
